pyproject.toml
[project]
name = "osia-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.68.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
    "uvicorn>=0.42.0",
    "websockets>=16.0",
]
